refactor(questions): merge reopen and resolve cmds and add new commendation logic
